+++ M2TWEOP-luaPlugin/luaPlugin/gameHelpers.h
@@ -34,7 +34,7 @@ namespace gameHelpers
 	int getMercUnitNum(mercPool* mercPool);
 	mercPoolUnit* getMercUnit(mercPool* pool, int index);
 	void setMercReligion(mercPoolUnit* unit, int religion, bool set);
-	mercPoolUnit* addMercUnit(mercPool* mercPool, int idx, int exp, int cost, float repmin, float repmax, int maxunits, float startpool, int startyear, int endyear, int crusading);
+	mercPoolUnit* addMercUnit(mercPool* mercPool, int idx, int exp, int cost, float repmin, float repmax, int maxunits, float startpool, float startyear, float endyear, int crusading);
 	void saveGame(const char* path);
 };
